+++ audionimbus/src/baking/pathing.rs
@@ -187,7 +187,7 @@ pub struct PathBakeParams {
 pub mod tests {
     use crate::*;
 
-    fn test_scene(context: &Context) -> Scene<DefaultRayTracer> {
+    fn test_scene(context: &Context) -> Scene<'_, DefaultRayTracer> {
         let mut scene = Scene::try_new(context).unwrap();
 
         // Create a simple room mesh.