+++ audionimbus/src/geometry/instanced_mesh.rs
@@ -69,7 +69,7 @@ unsafe impl Send for InstancedMesh {}
 #[derive(Debug, Clone)]
 pub struct InstancedMeshSettings<'a> {
     /// Handle to the scene to be instantiated.
-    pub sub_scene: &'a Scene,
+    pub sub_scene: &'a Scene<'a>,
 
     /// Local-to-world transform that places the instance within the parent scene.
     pub transform: Matrix<f32, 4, 4>,